change input type for unit price, grn table data update during table add isAcountPosting status logic in po and grn table when delete and update operation occured

+++ src/components/GoodsReceiveNoteInformation/Index/GRNInformationTable/GRNInfoList.js
@@ -26,13 +26,17 @@ import { useGetAllPurchaseOrderInformationQuery } from "../../../../redux/featur
 const GRNInfoList = ({ permission }) => {
   const [filterText, setFilterText] = useState("");
   const [resetPaginationToggle, setResetPaginationToggle] = useState(false);
-  const { data: grnAllInformation,isLoading:isGRNLoading, refetch } =
-    useGetAllGRNInformationQuery(undefined);
+  const {
+    data: grnAllInformation,
+    isLoading: isGRNLoading,
+    refetch,
+  } = useGetAllGRNInformationQuery(undefined);
   const { data: supplierInfo } = useGetAllSupplierInformationQuery(undefined);
   const { data: companyinfo } = useGetCompanyInfoQuery(undefined);
-  const {data:purchaseInfoData}=useGetAllPurchaseOrderInformationQuery(undefined)
+  const { data: purchaseInfoData } =
+    useGetAllPurchaseOrderInformationQuery(undefined);
   const [deleteGRNInfo] = useDeleteGRNInformationMutation();
-  const [selectSupplierPoNo, setSelectSupplierPoNo] = useState('');
+  const [selectSupplierPoNo, setSelectSupplierPoNo] = useState("");
   const [isTableDispaly, setIsTableDisplay] = useState(false);
   const [fromDate, setFromDate] = useState(
     new Date().toLocaleDateString("en-CA")
@@ -40,7 +44,7 @@ const GRNInfoList = ({ permission }) => {
   const [toDate, setToDate] = useState(new Date().toLocaleDateString("en-CA"));
   const [pOOptionsData, setPOOptionsData] = useState([]);
   const [filteredData, setFilteredData] = useState([]);
-  const [isFetchAfterDeleteData,setIsFetchAfterDeleteData]=useState(false)
+  const [isFetchAfterDeleteData, setIsFetchAfterDeleteData] = useState(false);
   const reportTitle = "Goods Receive Note";
 
   useEffect(() => {
@@ -56,15 +60,17 @@ const GRNInfoList = ({ permission }) => {
     };
     const poOptions = createPODropdown(purchaseInfoData);
     setPOOptionsData(poOptions);
-
   }, [purchaseInfoData]);
 
-  useEffect(()=>{
-    if(isFetchAfterDeleteData){
-      handleFilter()
+  useEffect(() => {
+    if (isFetchAfterDeleteData) {
+      setFilteredData(grnAllInformation);
+      handleFilter();
+      setIsFetchAfterDeleteData(false);
     }
-    setIsFetchAfterDeleteData(false)
-  },[isFetchAfterDeleteData])
+  }, [isFetchAfterDeleteData, grnAllInformation]);
+
+  console.log(filteredData);
 
   const columns = [
     {
@@ -122,7 +128,7 @@ const GRNInfoList = ({ permission }) => {
       sortable: true,
       center: true,
       filterable: true,
-      width:'210px'
+      width: "210px",
     },
 
     {
@@ -186,36 +192,45 @@ const GRNInfoList = ({ permission }) => {
                 marginLeft: "10px",
               }}
               onClick={() => {
-                swal({
-                  title: "Are you sure to delete this item?",
-                  text: "If once deleted, this item will not recovery.",
-                  icon: "warning",
-                  buttons: true,
-                  dangerMode: true,
-                }).then(async (willDelete) => {
-                  if (willDelete) {
-                    const response = await deleteGRNInfo(
-                      filteredData?._id
-                    ).unwrap();
-                    console.log(response);
-                    if (response.status === 200) {
-                      swal("Deleted!", "Your selected item has been deleted!", {
-                        icon: "success",
-                      });
-                      refetch();
-                      
-                      setIsFetchAfterDeleteData(true)
+                if (filteredData.isAccountPostingStatus === true) {
+                  swal(
+                    "Can not Update!",
+                    "Because already posted in accounting system",
+                    "error"
+                  );
+                }else{
+                  swal({
+                    title: "Are you sure to delete this item?",
+                    text: "If once deleted, this item will not recovery.",
+                    icon: "warning",
+                    buttons: true,
+                    dangerMode: true,
+                  }).then(async (willDelete) => {
+                    if (willDelete) {
+                      const response = await deleteGRNInfo(
+                        filteredData?._id
+                      ).unwrap();
+                      console.log(response);
+                      if (response.status === 200) {
+                        swal("Deleted!", "Your selected item has been deleted!", {
+                          icon: "success",
+                        });
+                        await refetch();
+  
+                        setIsFetchAfterDeleteData(true);
+                      } else {
+                        swal(
+                          "Error",
+                          "An error occurred while creating the data",
+                          "error"
+                        );
+                      }
                     } else {
-                      swal(
-                        "Error",
-                        "An error occurred while creating the data",
-                        "error"
-                      );
+                      swal(" Cancel! Your selected item is safe!");
                     }
-                  } else {
-                    swal(" Cancel! Your selected item is safe!");
-                  }
-                });
+                  });
+                }
+                
               }}
             >
               <FontAwesomeIcon icon={faTrash}></FontAwesomeIcon>
@@ -324,50 +339,45 @@ const GRNInfoList = ({ permission }) => {
   }, [filterText, resetPaginationToggle, companyinfo, reportTitle]);
 
   const handleFilter = async () => {
-if(isGRNLoading) {
-  console.log('loading')
-}
-else{
-  let filtered = grnAllInformation
-  if (fromDate && toDate && selectSupplierPoNo !=='') {
-    // Filter by date range and supplier PO number
-    filtered = await grnAllInformation.filter(item => {
-      const itemDate = new Date(item.receiveDate);
-      const isDateInRange =
-        itemDate >= new Date(fromDate) && itemDate <= new Date(toDate);
-      const isPonumberMatch = item.supplierPoNo === selectSupplierPoNo;
-      return isDateInRange && isPonumberMatch;
-    });
-    setFilteredData(filtered);
-  } else if (fromDate && toDate) {
-    filtered =await grnAllInformation?.filter(item => {
-      const itemDate = new Date(item.receiveDate);
-      return (
-        itemDate >= new Date(fromDate) && itemDate <= new Date(toDate)
-      );
-    });
-    setFilteredData(filtered);
-  } else if (selectSupplierPoNo) {
-    filtered = await grnAllInformation.filter(item => item.supplierPoNo === selectSupplierPoNo);
-  }
+    if (isGRNLoading) {
+      console.log("loading");
+    } else {
+      let filtered = grnAllInformation;
+      if (fromDate && toDate && selectSupplierPoNo !== "") {
+        // Filter by date range and supplier PO number
+        filtered = await grnAllInformation.filter((item) => {
+          const itemDate = new Date(item.receiveDate);
+          const isDateInRange =
+            itemDate >= new Date(fromDate) && itemDate <= new Date(toDate);
+          const isPonumberMatch = item.supplierPoNo === selectSupplierPoNo;
+          return isDateInRange && isPonumberMatch;
+        });
+      } else if (fromDate && toDate) {
+        filtered = await grnAllInformation?.filter((item) => {
+          const itemDate = new Date(item.receiveDate);
+          return itemDate >= new Date(fromDate) && itemDate <= new Date(toDate);
+        });
+      } else if (selectSupplierPoNo) {
+        filtered = await grnAllInformation.filter(
+          (item) => item.supplierPoNo === selectSupplierPoNo
+        );
+      }
 
-  // setFilteredData(filtered);
-  if(filtered?.length !==0)
-  {
-    setIsTableDisplay(true)
-  }
-  else{
-    setIsTableDisplay(false)
-    swal({
-      title: "Sorry!",
-      text: "No Data Available.",
-      icon: "warning",
-      button: "OK",
-    });
-  }
-  // setFilteredData(filtered);
-}
- 
+      setFilteredData(filtered);
+      if (filtered?.length !== 0) {
+        setIsTableDisplay(true);
+      } else {
+        setIsTableDisplay(false);
+        swal({
+          title: "Sorry!",
+          text: "No Data Available.",
+          icon: "warning",
+          button: "OK",
+        });
+      }
+      // setFilteredData(filtered);
+      refetch();
+    }
   };
 
   return (
@@ -380,154 +390,150 @@ else{
         }}
       >
         <div>
-        <h3 className="fw-bold mt-2">Goods Receive Note (GRN) List </h3>
-        <hr />
-        <div className="d-flex justify-content-between align-items-center w-75">
-          <div className="w-25">
-            <label htmlFor="">From Date</label>
-            <br />
-            <DatePicker
-              dateFormat="y-MM-dd"
-              className="text-center custom-datepicker2 "
-              calendarClassName="custom-calendar2"
-              selected={fromDate}
-              required
-              onChange={(fromDate) => {
-                console.log(fromDate);
-                if (fromDate > new Date()) {
-                  swal({
-                    title: "Select Valid Date",
-                    text: "Date should be equal or earlier than today",
-                    icon: "warning",
-                    button: "OK",
-                  });
-                } else {
-                  setFromDate(fromDate.toLocaleDateString("en-CA"));
-                }
-              }}
-            />
-          </div>
-          <div className="w-25">
-            <label htmlFor="">To Date</label>
-            <br />
-            <DatePicker
-              dateFormat="y-MM-dd"
-              className="text-center custom-datepicker2 "
-              calendarClassName="custom-calendar2"
-              selected={toDate}
-              required
-              onChange={(toDate) => {
-                console.log(toDate);
-                if (toDate > new Date()) {
-                  swal({
-                    title: "Select Valid Date",
-                    text: "Date should be equal or earlier than today",
-                    icon: "warning",
-                    button: "OK",
-                  });
-                } else {
-                  setToDate(toDate.toLocaleDateString("en-CA"));
-                }
-              }}
-            />
-          </div>
-          <div style={{ width: "35%" }}>
-            <div className="w-100">
-              <label htmlFor="">Supplier PONo</label>
+          <h3 className="fw-bold mt-2">Goods Receive Note (GRN) List</h3>
+          <hr />
+          <div className="d-flex justify-content-between align-items-center w-75">
+            <div className="w-25">
+              <label htmlFor="">From Date</label>
               <br />
+              <DatePicker
+                dateFormat="y-MM-dd"
+                className="text-center custom-datepicker2 "
+                calendarClassName="custom-calendar2"
+                selected={fromDate}
+                required
+                onChange={(fromDate) => {
+                  console.log(fromDate);
+                  if (fromDate > new Date()) {
+                    swal({
+                      title: "Select Valid Date",
+                      text: "Date should be equal or earlier than today",
+                      icon: "warning",
+                      button: "OK",
+                    });
+                  } else {
+                    setFromDate(fromDate.toLocaleDateString("en-CA"));
+                  }
+                }}
+              />
+            </div>
+            <div className="w-25">
+              <label htmlFor="">To Date</label>
+              <br />
+              <DatePicker
+                dateFormat="y-MM-dd"
+                className="text-center custom-datepicker2 "
+                calendarClassName="custom-calendar2"
+                selected={toDate}
+                required
+                onChange={(toDate) => {
+                  console.log(toDate);
+                  if (toDate > new Date()) {
+                    swal({
+                      title: "Select Valid Date",
+                      text: "Date should be equal or earlier than today",
+                      icon: "warning",
+                      button: "OK",
+                    });
+                  } else {
+                    setToDate(toDate.toLocaleDateString("en-CA"));
+                  }
+                }}
+              />
+            </div>
+            <div style={{ width: "35%" }}>
               <div className="w-100">
-                <Select
-                  class="form-select"
-                  className="w-100"
-                  aria-label="Default select example"
-                  name="poinfo"
-                  options={pOOptionsData}
-                  defaultValue={{
-                    label: "Select Supplier PONo",
-                    value: 0,
-                  }}
-                  value={pOOptionsData.filter(function (
-                    option
-                  ) {
-                    return (
-                      option.value === selectSupplierPoNo
-                    );
-                  })}
-                  styles={{
-                    control: (baseStyles, state) => ({
-                      ...baseStyles,
-                      width: "100%",
-                      borderColor: state.isFocused ? "#fff" : "#fff",
-                      border: "1px solid #2DDC1B",
-                    }),
-                    menu: (provided) => ({
-                      ...provided,
-                      zIndex: 9999,
-                      height: "auto",
-                      // overflowY: "scroll",
-                    }),
-                  }}
-                  theme={(theme) => ({
-                    ...theme,
-                    colors: {
-                      ...theme.colors,
-                      primary25: "#B8FEB3",
-                      primary: "#2DDC1B",
-                    },
-                  })}
-                  onChange={(e) => {
-                    setSelectSupplierPoNo(e.value);
-                  }}
-                ></Select>
+                <label htmlFor="">Supplier PONo</label>
+                <br />
+                <div className="w-100">
+                  <Select
+                    class="form-select"
+                    className="w-100"
+                    aria-label="Default select example"
+                    name="poinfo"
+                    options={pOOptionsData}
+                    defaultValue={{
+                      label: "Select Supplier PONo",
+                      value: 0,
+                    }}
+                    value={pOOptionsData.filter(function (option) {
+                      return option.value === selectSupplierPoNo;
+                    })}
+                    styles={{
+                      control: (baseStyles, state) => ({
+                        ...baseStyles,
+                        width: "100%",
+                        borderColor: state.isFocused ? "#fff" : "#fff",
+                        border: "1px solid #2DDC1B",
+                      }),
+                      menu: (provided) => ({
+                        ...provided,
+                        zIndex: 9999,
+                        height: "auto",
+                        // overflowY: "scroll",
+                      }),
+                    }}
+                    theme={(theme) => ({
+                      ...theme,
+                      colors: {
+                        ...theme.colors,
+                        primary25: "#B8FEB3",
+                        primary: "#2DDC1B",
+                      },
+                    })}
+                    onChange={(e) => {
+                      setSelectSupplierPoNo(e.value);
+                    }}
+                  ></Select>
+                </div>
               </div>
             </div>
-          </div>
-          <div>
-            <button
-              className="border-0 "
-              style={{
-                backgroundColor: "#2DDC1B",
-                color: "white",
-                padding: "5px 10px",
-                fontSize: "14px",
-                borderRadius: "5px",
-                width: "100px",
-                height: "38px",
-                marginLeft: "10px",
-                marginTop: "25px",
-              }}
-              onClick={handleFilter}
-            >
-              Show
-            </button>
-          </div>
-          <div>
-            <button
-              className="border-0 "
-              style={{
-                backgroundColor: "red",
-                color: "white",
-                padding: "5px 10px",
-                fontSize: "14px",
-                borderRadius: "5px",
-                width: "100px",
-                height: "38px",
-                marginLeft: "10px",
-                marginTop: "25px",
-              }}
-              onClick={() => {
-                setFromDate(new Date().toLocaleDateString('en-CA'));
-                setToDate(new Date().toLocaleDateString('en-CA'));
-                setSelectSupplierPoNo("");
-                setIsTableDisplay(false);
-              }}
-            >
-              Clear
-            </button>
+            <div>
+              <button
+                className="border-0 "
+                style={{
+                  backgroundColor: "#2DDC1B",
+                  color: "white",
+                  padding: "5px 10px",
+                  fontSize: "14px",
+                  borderRadius: "5px",
+                  width: "100px",
+                  height: "38px",
+                  marginLeft: "10px",
+                  marginTop: "25px",
+                }}
+                onClick={handleFilter}
+              >
+                Show
+              </button>
+            </div>
+            <div>
+              <button
+                className="border-0 "
+                style={{
+                  backgroundColor: "red",
+                  color: "white",
+                  padding: "5px 10px",
+                  fontSize: "14px",
+                  borderRadius: "5px",
+                  width: "100px",
+                  height: "38px",
+                  marginLeft: "10px",
+                  marginTop: "25px",
+                }}
+                onClick={() => {
+                  setFromDate(new Date().toLocaleDateString("en-CA"));
+                  setToDate(new Date().toLocaleDateString("en-CA"));
+                  setSelectSupplierPoNo("");
+                  setIsTableDisplay(false);
+                }}
+              >
+                Clear
+              </button>
+            </div>
           </div>
         </div>
-        </div>
-       
+
         {isTableDispaly ? (
           <div className=" ">
             <DataTable
